# openapi.yaml
...
swagger: "2.0"
info:
  description: Zylitics front end api
  title: Zylitics front end api
  # This version no. is just to showing in the console and has no affect on functionality, although it should be
  # updated whenever production version is updated.
  version: 0.3.9
host: front.api.zylitics.io
...
basePath: /beta
schemes:
  - http # TLS is terminated before hitting ESP, so only http.
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users:
    post:
      description: New user
      operationId: newUser
      parameters:
        - in: body
          name: newUserRequest
          required: true
          schema:
            $ref: '#/definitions/NewUserRequest'
      responses:
        200:
          description: NewUserResponse with User and customToken
          schema:
            $ref: '#/definitions/NewUserResponse'
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /users/current:
    get:
      description: get currently logged-in user
      operationId: getUser
      responses:
        200:
          description: The currently logged-in user
          schema:
            $ref: '#/definitions/User'
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /users/current/getUserPlan:
    get:
      description: get currently logged-in user's plan
      operationId: getUserPlan
      responses:
        200:
          description: The currently logged-in user's plan
          schema:
            $ref: '#/definitions/UsersPlan'
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
  /users/current/updateUserProfile:
    patch:
      description: update currently logged-in user's profile
      operationId: updateUserProfile
      parameters:
        - in: body
          name: userUpdatableProfile
          required: true
          schema:
            $ref: '#/definitions/UserUpdatableProfile'
      responses:
        200:
          description: OK
        default:
          description: Unexpected exception
          schema:
            $ref: '#/definitions/ApiError'
...
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/zl-front-end"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "zl-front-end"
security:
  - firebase: []
...
definitions:
...
  NewUserRequest:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
        format: password
      timezone:
        type: string
      utcOffsetInMinutes:
        type: integer
      emailVerificationId:
        type: integer
      organizationName:
        type: string
  Organization:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  UsersPlan:
    type: object
    properties:
      planType:
        type: string
        enum: [FREE, PAID]
      planName:
        type: string
      displayName:
        type: string
      consumedMinutes:
        type: integer
      totalParallel:
        type: integer
      totalMinutes:
        type: integer
      billingCycleStart:
        type: integer
        format: int64
      billingCyclePlannedEnd:
        type: integer
        format: int64
  User:
    type: object
    properties:
      id:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      timezone:
        type: string
      role:
        type: string
        enum: [ADMIN, DEVELOPER]
      organizationId:
        type: integer
      shotBucketSessionStorage:
        type: string
      emailVerificationId:
        type: integer
        format: int64
      organization:
        $ref: '#/definitions/Organization'
      usersPlan:
        $ref: '#/definitions/UsersPlan'
  NewUserResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/User'
      customToken:
        type: string
  UserUpdatableProfile:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      timezone:
        type: string
      organizationName:
        type: string
...
  ApiError:
    type: object
    required:
      - message
    properties:
      message:
        type: string
...